Add —side flag to explicitly specify modrinth side
Modrinth allows mods to specify a “side”: whether they are required on the client, the server, or both. However, some mods don’t specify this, and currently packwiz fails for those mods. This commit adds a —side flag that allows the user to explicitly set whether the mod should be installed on the client, the server, or both.